It seems that the tests have been broken because of the modification of the domain decomposition scheme. The committed revision has an updated domain decomposition scheme as in commit 5f9dccac00ca51325898b508ed5de81a83386724.
In addition to that, member 'UserProcessParallelNodeDistributionList' has been added to the mesh class.
'UserProcessParallelNodeDistributionList' is a pointer to a user-defined function that when defined is called right after the new domain decomposition is determined.
That is the mechanism that allows modifying the decomposition by a "user."
Example of the practical application is the implementation of the periodic boundary conditions in AMPS where this mechanism is used to enforce 'real' and 'ghost' blocks be assigned to the same MPI process.