The class can be used to create a log file of a running process by forking a logger fork().
After the termination of the parent process, the logger is still running.
The logger saves the log file and then terminates itself.

Data exchange between the parent process and the logger:
1. Shared memory is used: ftok()/shmget()
2. The parent process does initialization of the shared memory
3. Synchronization between the parent and the forked logger is done with a semaphore

The log is created when
1. The parent process is terminated
2. Execution of a function exceeds a user-defined limit

The logger is a template class -> function-depended data can be saved in the logger at the logging points.

Need to be implemented: creating log files in different directories to distinguish between:
1. Unexpected termination of the parent process
2. Hanging the parent process
3. User-triggered termination of the parent process
4. User-requested creating the log file