Modified PIC::FieldSolver::Electromagnetic::ECSIM::UpdateJMassMatrix():
Implemented an additional mechanism for storing calculated stuff in the corner node's without race condition when calculated with multiple threads (std::threads or OpenMP). The mechanism is activated when _PIC_UPDATE_JMASS_MATRIX__MPI_MULTITHREAD_ == _PIC_MODE_ON_. In all other cases, the previously implemented routine is used.

The default number of std::threads used in calculating the mass matrix is thread_id_table_size=4. Need to experiment with that number.

The procedure requires 8 synchronizations + MPI_Allreduce to check whether the mesh was changed.

The code is fully functioning but the acceleration due to using std::threads very questionable. The time used on the barrier is minimal. Performance issues must be related to executing the threads (memory access, affinity?)

The algorithm would be good for CUDA, OpenACC, or OpenMP because there are no race conditions when accessing the state vector of the corner nodes. No locks are used.

Important: the algorithm works only with a uniform mesh!