Config.pl: 1. added parameter -test-blocks that defined the number of the test blocks for the concurrent execution of the tests 2. when the script is called with parameters '-install' or '-test-blocks', Config.pl compiles utility/TestScripts/table_parser.cpp and created the Makefile.test
MakefileTest/Table:
the new keyword 'Type=[amps,swmf]` is introduced to
combine all SWMF-related tests in a single target. The default value
assumed by table_parser.cpp is Type=amps. The point is that by combining
all SWMF-related in a single target, it is guaranteed that the same SWMF
installation is not used by several tests concurrently.