The size of the "next" and "prev" fields is set to 4 bytes each. So, the basic content of the state vector (1-byte descriptor, 8 bytes for next/prev, and 48 bytes for x and v) would fit into a single cache line. Added the capability to align the particle state vector to 64 to guarantee that the basic particle data is actually located in a single cache line.
Config.pl: added '-align', '-no-avx-mover', '-no-avx-matmul'

src/general/global.dfn: defined '_AVX_PARTICLE_MOVER_', '_AVX_MATMUL_',
and '_ALIGN_STATE_VECTORS_'

src/pic/pic.h: Set/Set of Next/Prev are unjested for 'int32_t' for the
appropriate fields

src/pic/picParticleDataMacro.h: size of 'next' and 'prev' are set to
sizeof(int32_t)

src/pic/pic_field_solver_ecsim.cpp: a few optimixation for when using
AVX.  'MassMatrix_GGD' is alightn to 64 and the size is changed to
MassMatrix_GGD[8][8][12] (when it is actually needed only [8][8][9]) -
that is needed to have the beginning of MassMatrix_GGD[:][:][12] align
to 32, which is needed to AVX2)

src/pic/pic_mover_boris.cpp: AVX can be turned off
PIC::Mover::Lapenta2017() even when it is used in the rest of the code

src/pic/pic_pbuffer.cpp: When _CUDA_MODE_==_OFF_ and
_ALIGN_STATE_VECTORS_==_ON_ in the function
PIC::ParticleBuffer::Init() - 1) the size of the particle state vector
is align with 64 byte and 2) that particle buffer is allcated with
aligned_alloc(64,..) to force the beginning of a particle state vector
starts at the beginning of a cache line.

srcInterface/LinearSystemCornerNode.h: AVX can be turned off in the
matrix/vector multiplication function even when AVX is used in the rest
of the code.