The structire of the first byte of the particle state vector is changed. In the current implementation:
bits 0-5: particle species index
bit    6: flag that models can used to set the initialization status.
Currently the flag will by used only by the gyrokinetic model to
determine whether the particles' magnetic moment is initialized. The
flag can be set/tested with functions SetInitFlag() and TestInitFlag()
bit    7: the status flag determening whether the particle is "used
currently in the simulation" or "available to be added in the
simulation"