logger is modified: Now: The mechanism to determine that the run is terminated does not rely ONLY on checking the existence of the process with that parent process PID.
Checking the existence of the parent process
required starting a daemon detached from the run that
is affected by killing the parent process. Implemented procedure for
starting a daemon works under Red Hat, but not SUSE.

NOW: in addition to
checking the existence of the parent processes, the logger check
existence of file "AMPS_LOGGER_TERMINATE" that signals termination of
the run. If the file is found, the logger prints the log is printed and
terminates itself.

The logger creates the file "AMPS_LOGGER_TERMINATE" when the code sits
in a  function time that exceeds a predetermined limit. In such a case:
A log is created.
"AMPS_LOGGER_TERMINATE" is created.
Logger waits for 10 sec.
Logger terminates the parent process.
The logger's Server() is called 1/10 sec, so 10 sec of waiting time is
sufficient.

The best part - it works on Pleiades.