1. added: utility/compare_runs.cpp - the utility compare the contect of the shared memory populated by two concurrently executed runs 2. PIC::Debugger::ConcurrentDebug::NewEntry() is added in SetV/SetX when _PIC__DEBUG_CONCURRENT_RUNS_ == _PIC_MODE_ON_
3. There is a sincronization error somewhere in the populating/comparing
the content of the shared memory. As of now, two semaphores are used:
(1) sem_data - to exclude the concurrent writting and reading to/from shared
memorym, and (2) sem_exit - to allows AMPS' processes exit from
PIC::Debugger::ConcurrentDebug::NewEntry(). In the latter case, sem_wait
is called by MPI processes of AMPS, and sem_post by the comparing
utility, after comparing of the shared memory regions is completed.

!!!! There is a sincronization issue somewhere.