modified PIC::Parallel::ExchangeParticleData(): renamed to PIC::Parallel::ExchangeParticleData_unbuffered() added: PIC::Parallel::ExchangeParticleData_buffered()
PIC::Parallel::ExchangeParticleData_buffered -> collects the particle
data to be send in buffers, send them, and then unpack them
PIC::Parallel::ExchangeParticleData_unbuffered -> creates the MPI
type that contained pointers to the particles that needs to be send. On
the recieving side, the MPI type
contains pointers to the state vector of the newly arrived particles.
There is an issue: the function does not work with when OpenMP is used -
the function dies somewhere in
MPI calls.

new function PIC::Parallel::ExchangeParticleData() is just an interface
to call ExchangeParticleData_buffered() or
PIC::Parallel::ExchangeParticleData_unbuffered()

Which function will be called is defiend with macro
_PIC_PARTICLE_EXCHANGE_. The defauilt value of the macro
_PIC_PARTICLE_EXCHANGE_UNBUFFERED_